requirements.txt
Flask==2.2.3
streamlit==1.18.1
requests==2.28.2
python-dotenv==0.20.0
sentence-transformers==2.2.0
